+++ assets/javascripts/kitten/components/layout/marger.test.js
@@ -68,7 +68,7 @@ describe('<Marger />', () => {
         const styles = cleanStyles(marger.props().style)
 
         expect(styles).toEqual([
-          { marginTop: '2.5rem' },
+          { '@media (min-width: 0)': { marginTop: '2.5rem' } },
           { '@media (min-width: 480px)': { marginTop: '7.5rem' } },
           { '@media (min-width: 768px)': { marginTop: '1.25rem' } },
         ])
@@ -87,7 +87,9 @@ describe('<Marger />', () => {
           .first()
         const styles = cleanStyles(marger.props().style)
 
-        expect(styles).toEqual([{ marginTop: '2.5rem' }])
+        expect(styles).toEqual([
+          { '@media (min-width: 0)': { marginTop: '2.5rem' } },
+        ])
       })
 
       it('is overrided by the default value', () => {
@@ -101,7 +103,9 @@ describe('<Marger />', () => {
           .first()
         const styles = cleanStyles(marger.props().style)
 
-        expect(styles).toEqual([{ marginTop: '7.5rem' }])
+        expect(styles).toEqual([
+          { '@media (min-width: 0)': { marginTop: '7.5rem' } },
+        ])
       })
     })
   })
@@ -152,7 +156,7 @@ describe('<Marger />', () => {
         const styles = cleanStyles(marger.props().style)
 
         expect(styles).toEqual([
-          { marginBottom: '2.5rem' },
+          { '@media (min-width: 0)': { marginBottom: '2.5rem' } },
           { '@media (min-width: 480px)': { marginBottom: '7.5rem' } },
           { '@media (min-width: 768px)': { marginBottom: '1.25rem' } },
         ])
@@ -171,7 +175,9 @@ describe('<Marger />', () => {
           .first()
         const styles = cleanStyles(marger.props().style)
 
-        expect(styles).toEqual([{ marginBottom: '2.5rem' }])
+        expect(styles).toEqual([
+          { '@media (min-width: 0)': { marginBottom: '2.5rem' } },
+        ])
       })
 
       it('is overrided by the default value', () => {
@@ -185,7 +191,9 @@ describe('<Marger />', () => {
           .first()
         const styles = cleanStyles(marger.props().style)
 
-        expect(styles).toEqual([{ marginBottom: '7.5rem' }])
+        expect(styles).toEqual([
+          { '@media (min-width: 0)': { marginBottom: '7.5rem' } },
+        ])
       })
     })
   })