+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `k-TextInput--fiveDigits` class.
+
 ## [23.16.1] - 2018-11-30
 
 Fix: