+++ CHANGELOG.md
@@ -7,6 +7,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Add `links` props on `TeamCard` component.
   `socialLink` is deprecated.
 
+## [23.16.1] - 2018-11-30
+
+Fix:
+- Use media-query for default value on `Marger` component when prop `top`
+  or `bottom` is an object.
+
 ## [23.16.0] - 2018-11-30
 
 Feature: